Don't clear afs_stats_cmperf too early during shutdown
Commit 21cbf7fee0a089d94f62baa7df2422e7bc8293f7 activated some
previously unused cleanup code.  Part of this clears afs_stats_cmperf,
but these stats are used later to check for leaks of chunks allocated
with afs_AllocSmallSpace and afs_AllocLargeSpace.  The result is a
message about unfreed blocks in the syslog with negative counts.

Since the structure is already cleared later in the shutdown process,
just remove this instance.

Change-Id: I6bf2ab37752d2623d072469fb87fa6d06e85b9ec
Reviewed-on: http://gerrit.openafs.org/1229
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>