OpenBSD: move AFS memory type offet
The internal malloc memory types for IPv6 (M_IP6OPT, ...) conflict
with the type numbers used to designate AFS memory (M_AFSFID, etc.).
This change moves the AFS memory type to a new number that does not
conflict. This is not a serious issue but can create real confusion
when trying to debug or track memory issues, among other things.

Change-Id: Icccaa7c06443cab008013414cb3c2ab4ea08889e
Reviewed-on: http://gerrit.openafs.org/1271
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>