Add support for OpenBSD 4.6
Add config param header and sysname number for OpenBSD 4.6.

As well, add an additional parameter to a call to ifa_ifwithnet()
when looking up the MTU for an interface to indicate that the call
should use the default routing table. With the advent of OpenBSD
4.6, the system has started to make provisions for multiple routing
tables which included a change to the calling sequence for this
routine.

Change-Id: Ia616400094ed36bc6f0e37f86d3b7448d5bd32c2
(cherry picked from commit ab17ebef16a04180b2ede8c8ed07cd9c1871ae42)
Reviewed-on: http://gerrit.openafs.org/1261
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>