OpenBSD: eliminate use of VREF() macro
OpenBSD 4.7 (currently in beta) removes the (dummy) VREF() macro from the
system header files. For at least as far back as OpenBSD 3.6, all this
macro does is expand to a call to a system routine by the same name in
lower case. This patch therefore replaces the single OpenBSD use of
VREF() with a direct call to the routine to which it expanded.

Change-Id: I44888ed1ba7ba8e7150202798def10bed2ab3158
Reviewed-on: http://gerrit.openafs.org/1532
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>
(cherry picked from commit a31bbd3123dd7bd5c0f73c1bcd89548d489930a8)
Change-Id: I442d5e9cfa9bd7f5886692c0afcbb3d220d0ca5b
Reviewed-on: http://gerrit.openafs.org/1509