Windows: buffers whose offsets are beyond EOF should be zero filled and locally allocated
When a buffer is being allocated for an offset that exceeds the
file length as known to the file server, that buffer should be zero
filled and it does not require server validation.

Previously all buffers were populated with a FetchData call.
This is wasteful of time and server resources when there is a
valid callback registration because the serverLength is known
to be valid.

LICENSE MIT

Change-Id: I13f9a8dbd6387c3ba71638a682e995d7a7d0862f
Reviewed-on: http://gerrit.openafs.org/1664
Tested-by: Jeffrey Altman <jaltman@openafs.org>
Tested-by: Asanka Herath <asanka@secure-endpoints.com>
Reviewed-by: Asanka Herath <asanka@secure-endpoints.com>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>