Fix build for FBSD80
The change to a dynamically-allocated group list came before
the change to allow the maximum size of that group list to
be set as a tuneable at boot.  The 8.0 release happened to
come between them, so we must treat both cases.
(Note that AFS_FBSD81_ENV is not yet defined anywhere; that
will come later.)

Change-Id: I87a0e2cff3c42de60d512fe5653abe0161afc789
Reviewed-on: http://gerrit.openafs.org/1676
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>